+++ tests/Support/ClassEnum.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Html\Tests\Support;
+
+enum ClassEnum: string
+{
+    case TEST_CLASS_1 = 'test-class-1';
+    case TEST_CLASS_2 = 'test-class-2';
+}